rename renote tooltip component
The tooltip that is used for renotes can be used in other cases as well.
